// src/data/supabase.d.ts
export type Database = {
  public: {
    Tables: {
      sandwich_schedule: {
        Row: {
          date: string | null;
          days: string;
          dropdown_data: string | null;
          id: string;
          period: string | null;
          sandwich_type: SandwichType;
        };
        Insert: {
          date?: string | null;
          days: string;
          dropdown_data?: string | null;
          id: string;
          period?: string | null;
          sandwich_type: SandwichType;
        };
        Update: {
          date?: string | null;
          days?: string;
          dropdown_data?: string | null;
          id?: string;
          period?: string | null;
          sandwich_type?: SandwichType;
        };
        Relationships: [];
      };
      work_schedule: {
        Row: {
          date: string | null;
          id: number;
          workers: string[] | null;
        };
        Insert: {
          date?: string | null;
          id?: number;
          workers?: string[] | null;
        };
        Update: {
          date?: string | null;
          id?: number;
          workers?: string[] | null;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema['Tables'] & PublicSchema['Views'])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
        Database[PublicTableNameOrOptions['schema']]['Views'])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
      Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] & PublicSchema['Views'])
    ? (PublicSchema['Tables'] & PublicSchema['Views'])[PublicTableNameOrOptions] extends {
        Row: infer R;
      }
      ? R
      : never
    : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends keyof PublicSchema['Enums'] | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
    ? PublicSchema['Enums'][PublicEnumNameOrOptions]
    : never;